pyproject.toml
[project]
name = "cointrader"
version = "0.1.0"
description = "Burst-coin bull-flag trading bot for Coinbase"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "coinbase-advanced-py>=1.8.0",
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "rich>=13.9.0",
    "ta>=0.11.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "pyjwt>=2.10.1",
    "cryptography>=46.0.3",
    "pyarrow>=22.0.0",
    "duckdb>=1.4.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
